requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[argon2]>=1.7.4
python-multipart>=0.0.6
aiofiles>=23.0.0
psycopg2-binary>=2.9.9